requirements.txt
requests
beautifulsoup4
langchain
transformers
google-generativeai